[BUG FIX] Use ENDOFCHAIN value to test 1st Difat
The first Difat sector is identified within the compound file header (CFH)
by the member data named DifatSect1. Previously the test for this variable
was set at 0, which is was truly fool-hardy of 'yours truly'. This was
inspite of the existence of the ENDOFCHAIN constant which was set at
0xFFFFFFFE, in line with the format. Accordingly, this was changed and the
constants earlier defined in the implementation file were relocated to the
header along with the others that were defined there - it was this poor
location that made this issue difficult to spot in the first place.