Introduced DLL to project
In the course of building the GUI, it became apparent that there would be
repetitious use of source code across the project. Thus, it was decided
that dynamic/shared libraries would be used across the different projects.
This was done and relevant source file changes made as well. At this
point, all the key header files are lumped in one location - for the
purpose of experimentation. Later, a more intuitive directory structure
will be implemented.